Validating overlaps in ReceiverRecovery class also
The front end will now properly render errors. It's also stricter (in case the recovery is saved alone, without saving the deployment).
