+++ grails-app/domain/au/org/emii/aatams/ReceiverDeployment.groovy
@@ -135,32 +135,9 @@ class ReceiverDeployment
     }
 
     static def conflictingDeploymentValidator = { initialisationDateTime, deployment ->
-        // All non-null deployment intervals for receiver, other than 'this' deployment's.
-        def conflictingDeployments = deployment.receiver?.deployments.findAll {
-            !it.same(deployment) && it.undeployableInterval
-        }.findAll {
-            def checkInterval = it.undeployableInterval
-
-            def containedBy =
-                checkInterval.containsAny([
-                    deployment.initialisationDateTime,
-                    deployment.deploymentDateTime
-                ])
-
-            def overlappedBy = deployment.undeployableInterval?.overlaps(checkInterval)
-
-            return containedBy || overlappedBy
-        }
-
-        if (conflictingDeployments.isEmpty()) {
-            return true
-        }
-
-        [
-            'receiverDeployment.initialisationDateTime.conflictingDeployment',
-            deployment.receiver,
-            conflictingDeployments
-        ]
+        ReceiverDeploymentValidator.conflictingDeploymentValidator(
+            initialisationDateTime, deployment
+        )
     }
 
     // Don't want to override 'equals()' as this causes unexpected behaviour with GORM.
